Core layer refactoring
