+++ ChitChat02/CoreLayer/Networking/Parser/IParser.swift
@@ -27,24 +27,3 @@ struct AvatarItemApiModel: Codable {
         case id, author, url, downloadUrl = "download_url"
     }
 }
-
-class AvatarListParser: IParser {
-    typealias Model = [AvatarItemApiModel]
-    
-    func parse(data: Data) -> [AvatarItemApiModel]? {
-        do {
-            let avatarList = try JSONDecoder().decode([AvatarItemApiModel].self, from: data)
-            return avatarList
-        } catch {
-            fatalError(error.localizedDescription)
-        }
-    }
-}
-
-class ImageParser: IParser {
-    typealias Model = Data
-    
-    func parse(data: Data) -> Data? {
-        return data
-    }
-}
